replacing int with long

+++ src/main/java/ru/Netology/javaqa/ArrayOfMonthlySales/services/StatsService.java
@@ -4,66 +4,66 @@ public class StatsService {
 
     public int theAmountOfSales(long[] sales) {
         //сумма всех продаж
-        int total = 0;
-        for (int i = 0; i < sales.length; i++) {
-            total += sales[i];
+        long total = 0;
+        for (long i = 0; i < sales.length; i++) {
+            total += sales[(int) i];
         }
-        return total;
+        return (int) total;
     }
 
     public int averageSalesAmount(long[] sales) {
         //средняя сумма продаж в месяц
-        int total = 0;
-        for (int i = 0; i < sales.length; i++) {
-            total += sales[i];
+        long total = 0;
+        for (long i = 0; i < sales.length; i++) {
+            total += sales[(int) i];
         }
-        return total / sales.length;
+        return (int) (total / sales.length);
     }
 
     public int minSales(long[] sales) {
         //минимум продаж
-        int minMonth = 0;
-        for (int i = 0; i < sales.length; i++) {
-            if (sales[i] <= sales[minMonth]) {
+        long minMonth = 0;
+        for (long i = 0; i < sales.length; i++) {
+            if (sales[(int) i] <= sales[(int) minMonth]) {
                 minMonth = i;
             }
         }
-        return minMonth + 1;
+        return (int) (minMonth + 1);
     }
 
     public int maxSales(long[] sales) {
         // максимум продаж
-        int maxMonth = 0;
-        for (int i = 0; i < sales.length; i++) {
-            if (sales[i] >= sales[maxMonth]) {
+        long maxMonth = 0;
+        for (long i = 0; i < sales.length; i++) {
+            if (sales[(int) i] >= sales[(int) maxMonth]) {
                 maxMonth = i;
             }
         }
-        return maxMonth + 1;
+        return (int) (maxMonth + 1);
     }
 
     public int belowAverage(long[] sales) {
         //продажи ниже среднего
-        int numberOfMonth = 0;
-        int average = (int) averageSalesAmount (sales);
-        for (int i = 0; i < sales.length; i++) {
-            if (sales[i] >= average) {
+        long numberOfMonth = 0;
+        long average = (int) averageSalesAmount(sales);
+        for (long i = 0; i < sales.length; i++) {
+            if (sales[(int) i] >= average) {
                 numberOfMonth = numberOfMonth + 1;
             }
         }
-        return numberOfMonth;
+        return (int) numberOfMonth;
     }
 
     public int aboveAverage(long[] sales) {
         //продажи выше среднего
-        int numberOfMonth = 0;
-        int average = (int) averageSalesAmount (sales);
-        for (int i = 0; i < sales.length; i++) {
-            if (sales[i] <= average) {
+        long numberOfMonth = 0;
+        long average = (int) averageSalesAmount(sales);
+        for (long i = 0; i < sales.length; i++) {
+            if (sales[(int) i] <= average) {
                 numberOfMonth = numberOfMonth + 1;
             }
         }
-        return numberOfMonth;
+        return (int) numberOfMonth;
     }
 }
